Removes the string matching and extraction from TaskUtils. Fixes the commits wrt Identity and CassandraConfig removing options such that the null pointer exceptions they avoided are no longer thrown.
