+++ cassandra-commons/src/main/java/com/mesosphere/dcos/cassandra/common/config/CassandraConfig.java
@@ -322,7 +322,7 @@ public class CassandraConfig {
         this.cpus = cpus;
         this.memoryMb = memoryMb;
         this.diskMb = diskMb;
-        this.replaceIp = replaceIp;
+        this.replaceIp = (replaceIp != null) ? replaceIp : "";
         this.heap = heap;
         this.location = location;
         this.jmxPort = jmxPort;
