+++ cassandra-executor/src/main/java/com/mesosphere/dcos/cassandra/executor/tasks/UploadSnapshot.java
@@ -22,14 +22,18 @@ public class UploadSnapshot implements Runnable {
     private BackupUploadTask cassandraTask;
     private BackupStorageDriver backupStorageDriver;
 
-    public UploadSnapshot(ExecutorDriver driver, NodeProbe probe, CassandraTask cassandraTask, BackupStorageDriver backupStorageDriver) {
+    public UploadSnapshot(
+            ExecutorDriver driver,
+            NodeProbe probe,
+            CassandraTask cassandraTask,
+            String nodeId,
+            BackupStorageDriver backupStorageDriver) {
         this.probe = probe;
         this.driver = driver;
         this.cassandraTask = (BackupUploadTask) cassandraTask;
-        final int nodeId = TaskUtils.taskIdToNodeId(this.cassandraTask.getId());
         this.backupStorageDriver = backupStorageDriver;
         context = new BackupContext();
-        context.setNodeId(nodeId+"");
+        context.setNodeId(nodeId);
         context.setName(this.cassandraTask.getBackupName());
         context.setExternalLocation(this.cassandraTask.getExternalLocation());
         context.setS3AccessKey(this.cassandraTask.getS3AccessKey());
