+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/plan/backup/BackupSnapshotPhase.java
@@ -1,41 +1,40 @@
-package com.mesosphere.dcos.cassandra.scheduler.backup;
+package com.mesosphere.dcos.cassandra.scheduler.plan.backup;
 
-import com.mesosphere.dcos.cassandra.common.tasks.backup.RestoreContext;
+import com.mesosphere.dcos.cassandra.common.tasks.backup.BackupContext;
 import com.mesosphere.dcos.cassandra.scheduler.offer.ClusterTaskOfferRequirementProvider;
 import com.mesosphere.dcos.cassandra.scheduler.tasks.AbstractClusterTaskPhase;
 import com.mesosphere.dcos.cassandra.scheduler.tasks.CassandraTasks;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
 /**
- * During download snapshot phase, snapshotted data will be downloaded to all the cassandra node from
- * external location.
+ * During snapshot phase, data will be snapshotted across all cassandra nodes.
  */
-public class DownloadSnapshotPhase extends AbstractClusterTaskPhase<DownloadSnapshotBlock, RestoreContext> {
-    private static final Logger LOGGER =
-            LoggerFactory.getLogger(DownloadSnapshotPhase.class);
+public class BackupSnapshotPhase extends AbstractClusterTaskPhase<BackupSnapshotBlock, BackupContext> {
 
-    public DownloadSnapshotPhase(
-            RestoreContext context,
+    private static final Logger LOGGER =
+            LoggerFactory.getLogger(BackupSnapshotPhase.class);
+
+    public BackupSnapshotPhase(
+            BackupContext backupContext,
             CassandraTasks cassandraTasks,
             ClusterTaskOfferRequirementProvider provider) {
-        super(context, cassandraTasks, provider);
+        super(backupContext, cassandraTasks, provider);
     }
 
-    protected List<DownloadSnapshotBlock> createBlocks() {
+    protected List<BackupSnapshotBlock> createBlocks() {
         final List<String> daemons =
                 new ArrayList<>(cassandraTasks.getDaemons().keySet());
         Collections.sort(daemons);
-        return daemons.stream().map(daemon -> DownloadSnapshotBlock.create(
+        return daemons.stream().map(daemon -> BackupSnapshotBlock.create(
                 daemon,
                 cassandraTasks,
                 provider,
                 context
         )).collect(Collectors.toList());
     }
+
 }
