+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/plan/backup/DownloadSnapshotPhase.java
@@ -1,40 +1,41 @@
-package com.mesosphere.dcos.cassandra.scheduler.backup;
+package com.mesosphere.dcos.cassandra.scheduler.plan.backup;
 
-import com.mesosphere.dcos.cassandra.common.tasks.backup.BackupContext;
+import com.mesosphere.dcos.cassandra.common.tasks.backup.RestoreContext;
 import com.mesosphere.dcos.cassandra.scheduler.offer.ClusterTaskOfferRequirementProvider;
 import com.mesosphere.dcos.cassandra.scheduler.tasks.AbstractClusterTaskPhase;
 import com.mesosphere.dcos.cassandra.scheduler.tasks.CassandraTasks;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.stream.Collectors;
 
 /**
- * During snapshot phase, data will be snapshotted across all cassandra nodes.
+ * During download snapshot phase, snapshotted data will be downloaded to all the cassandra node from
+ * external location.
  */
-public class BackupSnapshotPhase extends AbstractClusterTaskPhase<BackupSnapshotBlock, BackupContext> {
-
+public class DownloadSnapshotPhase extends AbstractClusterTaskPhase<DownloadSnapshotBlock, RestoreContext> {
     private static final Logger LOGGER =
-            LoggerFactory.getLogger(BackupSnapshotPhase.class);
+            LoggerFactory.getLogger(DownloadSnapshotPhase.class);
 
-    public BackupSnapshotPhase(
-            BackupContext backupContext,
+    public DownloadSnapshotPhase(
+            RestoreContext context,
             CassandraTasks cassandraTasks,
             ClusterTaskOfferRequirementProvider provider) {
-        super(backupContext, cassandraTasks, provider);
+        super(context, cassandraTasks, provider);
     }
 
-    protected List<BackupSnapshotBlock> createBlocks() {
+    protected List<DownloadSnapshotBlock> createBlocks() {
         final List<String> daemons =
                 new ArrayList<>(cassandraTasks.getDaemons().keySet());
         Collections.sort(daemons);
-        return daemons.stream().map(daemon -> BackupSnapshotBlock.create(
+        return daemons.stream().map(daemon -> DownloadSnapshotBlock.create(
                 daemon,
                 cassandraTasks,
                 provider,
                 context
         )).collect(Collectors.toList());
     }
-
 }
