+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/plan/backup/UploadBackupPhase.java
@@ -1,4 +1,4 @@
-package com.mesosphere.dcos.cassandra.scheduler.backup;
+package com.mesosphere.dcos.cassandra.scheduler.plan.backup;
 
 import com.mesosphere.dcos.cassandra.common.tasks.backup.BackupContext;
 import com.mesosphere.dcos.cassandra.scheduler.offer.ClusterTaskOfferRequirementProvider;
