+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/resources/BackupResource.java
@@ -4,7 +4,7 @@ import com.codahale.metrics.annotation.Timed;
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.mesosphere.dcos.cassandra.common.tasks.backup.BackupContext;
-import com.mesosphere.dcos.cassandra.scheduler.backup.BackupManager;
+import com.mesosphere.dcos.cassandra.scheduler.plan.backup.BackupManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
