+++ cassandra-scheduler/src/main/java/com/mesosphere/dcos/cassandra/scheduler/resources/RestoreResource.java
@@ -3,7 +3,7 @@ package com.mesosphere.dcos.cassandra.scheduler.resources;
 import com.codahale.metrics.annotation.Timed;
 import com.google.inject.Inject;
 import com.mesosphere.dcos.cassandra.common.tasks.backup.RestoreContext;
-import com.mesosphere.dcos.cassandra.scheduler.backup.RestoreManager;
+import com.mesosphere.dcos.cassandra.scheduler.plan.backup.RestoreManager;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
